+++ rasa/cli/train.py
@@ -118,7 +118,14 @@ def train(args: argparse.Namespace) -> Optional[Text]:
         get_validated_path(f, "data", DEFAULT_DATA_PATH) for f in args.data
     ]
 
-    return rasa.train(domain, config, training_files, args.out, args.force)
+    return rasa.train(
+        domain,
+        config,
+        training_files,
+        args.out,
+        args.force,
+        extract_additional_arguments(args),
+    )
 
 
 def train_core(
@@ -143,7 +150,14 @@ def train_core(
 
         config = get_validated_path(args.config, "config", DEFAULT_CONFIG_PATH)
 
-        return train_core(args.domain, config, stories, output, train_path)
+        return train_core(
+            args.domain,
+            config,
+            stories,
+            output,
+            train_path,
+            extract_additional_arguments(args),
+        )
     else:
         from rasa.core.train import do_compare_training
 
@@ -162,3 +176,13 @@ def train_nlu(
     nlu_data = get_validated_path(args.nlu, "nlu", DEFAULT_DATA_PATH)
 
     return train_nlu(config, nlu_data, output, train_path)
+
+
+def extract_additional_arguments(args: argparse.Namespace) -> typing.Dict:
+    return {
+        "augmentation_factor": args.augmentation,
+        "dump_stories": args.dump_stories,
+        "debug_plots": args.debug_plots,
+        "percentages": args.percentages,
+        "runs": args.runs,
+    }