+++ rasa/cli/utils.py
@@ -1,4 +1,5 @@
 import os
+import sys
 from typing import Any, Callable, Dict, Optional, Text
 
 from rasa.constants import DEFAULT_MODELS_PATH
@@ -115,8 +116,8 @@ def minimal_kwargs(kwargs: Dict[Text, Any], func: Callable) -> Dict[Text, Any]:
     return {k: v for k, v in kwargs.items() if k in possible_arguments}
 
 
-def print_success(text: Text):
-    print_color(text, bcolors.OKGREEN)
+def print_success(*args: Any):
+    print_color(*args, color=bcolors.OKGREEN)
 
 
 class bcolors(object):
@@ -130,17 +131,22 @@ class bcolors(object):
     UNDERLINE = "\033[4m"
 
 
-def wrap_with_color(text: Text, color: Text):
-    return color + text + bcolors.ENDC
+def wrap_with_color(*args: Any, color: Text):
+    return color + " ".join(str(s) for s in args) + bcolors.ENDC
 
 
-def print_color(text: Text, color: Text):
-    print (wrap_with_color(text, color))
+def print_color(*args: Any, color: Text):
+    print (wrap_with_color(*args, color=color))
 
 
-def print_warning(text: Text):
-    print_color(text, bcolors.WARNING)
+def print_warning(*args: Any):
+    print_color(*args, color=bcolors.WARNING)
 
 
-def print_error(text: Text):
-    print_color(text, bcolors.FAIL)
+def print_error(*args: Any):
+    print_color(*args, color=bcolors.FAIL)
+
+
+def signal_handler(sig, frame):
+    print ("Goodbye 👋")
+    sys.exit(0)